Update ui data type to Float

+++ graphql/typeDefs/nodeUI.js
@@ -1,7 +1,7 @@
 const nodeUISchema = `
   type NodeUI {
-    x: Int!
-    y: Int!
+    x: Float!
+    y: Float!
   }
 `
